feat: Add global error view
