+++ packages/vitnode/package.json
@@ -6,16 +6,16 @@
     "@hono/zod-openapi": "^0.19.6",
     "@hono/zod-validator": "^0.5.0",
     "@hookform/resolvers": "^5.0.1",
-    "@react-email/components": "0.0.36",
-    "@swc/cli": "^0.7.3",
+    "@react-email/components": "0.0.37",
+    "@swc/cli": "0.6.0",
     "@swc/core": "^1.11.24",
-    "@testing-library/react": "^14.2.1",
-    "@vitejs/plugin-react": "^4.2.1",
-    "@vitest/coverage-v8": "^1.2.1",
+    "@testing-library/react": "^16.3.0",
     "@types/node": "^22.15.3",
     "@types/nodemailer": "^6.4.17",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
+    "@vitejs/plugin-react": "^4.4.1",
+    "@vitest/coverage-v8": "^3.1.2",
     "chokidar": "^4.0.3",
     "concurrently": "^9.1.2",
     "dotenv": "^16.5.0",
@@ -24,22 +24,22 @@
     "eslint": "^9.25.1",
     "eslint-config-typescript-vitnode": "workspace:*",
     "hono": "^4.7.8",
-    "jsdom": "^23.2.0",
+    "jsdom": "^26.1.0",
     "lucide-react": "^0.503.0",
     "next": "^15.3.1",
     "next-intl": "^4.1.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-email": "^4.0.7",
+    "react-email": "^4.0.11",
     "react-hook-form": "^7.56.1",
     "sonner": "^2.0.3",
-    "tailwindcss": "^4.1.4",
+    "tailwindcss": "^4.1.5",
     "tsc-alias": "^1.8.15",
     "tsup": "^8.4.0",
     "tw-animate-css": "^1.2.8",
     "typescript": "^5.8.3",
-    "vite-tsconfig-paths": "^4.3.1",
-    "vitest": "^1.2.1",
+    "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "^3.1.2",
     "zod": "^3.24.3"
   },
   "bin": {
@@ -62,7 +62,7 @@
     "build:scripts": "tsup",
     "start:scripts": "bun ./dist/scripts/scripts.js",
     "build": "tsc && swc src -d dist --config-file .swcrc && tsc-alias -p tsconfig.json",
-    "dev": "concurrently \"tsc -w --preserveWatchOutput\" \"swc src -d dist --config-file .swcrc -w\"",
+    "dev": "concurrently \"tsc -w --preserveWatchOutput\" \"swc src -d dist --config-file .swcrc -w\" \"tsc-alias -w\"",
     "dev:email": "email dev",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
@@ -73,7 +73,7 @@
   "type": "module",
   "dependencies": {
     "@dnd-kit/core": "^6.3.1",
-    "@tanstack/react-query": "^5.74.11",
+    "@tanstack/react-query": "^5.75.0",
     "cache-manager": "^6.4.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -84,7 +84,7 @@
     "postgres": "^3.4.5",
     "radix-ui": "^1.3.4",
     "react-scan": "^0.3.3",
-    "resend": "^4.4.1",
+    "resend": "^4.5.0",
     "tailwind-merge": "^3.2.0",
     "vaul": "^1.1.2"
   }
