+++ packages/vitnode/src/views/admin/layouts/sidebar/sidebar.tsx
@@ -7,7 +7,7 @@ import { NavSidebarAdmin } from './nav/nav';
 
 export const SidebarAdmin = () => {
   return (
-    <Sidebar collapsible="icon">
+    <Sidebar>
       <SidebarHeader className="flex h-16 flex-row items-center border-b">
         <Link className="px-2" href="/admin/core">
           <LogoVitNode className="size-8" small />
