+++ plugins/blog/package.json
@@ -32,7 +32,7 @@
     "zod": "^3.24.3"
   },
   "devDependencies": {
-    "@swc/cli": "^0.7.3",
+    "@swc/cli": "0.6.0",
     "@swc/core": "^1.11.24",
     "concurrently": "^9.1.2",
     "eslint": "^9.25.1",
